Improved thread stability in audio buffer playing. Still have issues with thread safety.

+++ ensemble/src/ensemble/apps/pd_testing/Pd_Audio_Buffer.java
@@ -2,17 +2,17 @@ package ensemble.apps.pd_testing;
 
 public class Pd_Audio_Buffer 
 {
-	private float[ ] audio_samples;
+	private byte[ ] audio_samples;
 	private int pd_time_tag;
 	private String source;
 	
-	public Pd_Audio_Buffer ( float[ ] new_audio_samples, int new_time_tag, String new_source )
+	public Pd_Audio_Buffer ( byte[ ] new_audio_samples, int new_time_tag, String new_source )
 	{
 		audio_samples = new_audio_samples;
 		pd_time_tag = new_time_tag;
 		source = new_source;
 	}
-	public float[ ] get_audio_samples ( )
+	public byte[ ] get_audio_samples ( )
 	{
 		return audio_samples;
 	}
